Proposed feature: make profile:build and profile:host optional
I've never written typescript before so this could use a check. But: I've tried to make it so that the profileCommand only adds build or host profile if the argument is defined. This is so that you can use the cmake generator for Conan in Conan 1.x (if you specify both, that particular generator breaks).